+++ src/main/java/com/evacipated/cardcrawl/modthespire/MTSClassPool.java
@@ -17,4 +17,9 @@ public class MTSClassPool extends ClassPool
     {
         return classLoader;
     }
+
+    public void setParent(ClassPool parent)
+    {
+        this.parent = parent;
+    }
 }